refactor: extract ml related files to folder and delete the unused ones
