--- app/core/recognition/index.ts
@@ -1,11 +1,0 @@
-export interface RecognitionResult {
-  inference: InferenceProbability
-  timestampStart: number,
-  timestampEnd: number,
-  recognitionTime: number,
-}
-
-export interface InferenceProbability {
-  class: string,
-  probability: number
-}
